Move shared heading travel math out of ion-storm module (#290).
Fleet trails need the same north-clockwise Δly helper; a neutral cartography module keeps that ownership clear.
