+++ packages/frontend/src/lib/cartography/ionStormMovement.ts
@@ -1,14 +1,4 @@
-/** Per-turn travel delta in game light-years (heading 0 = north, clockwise). */
-export function headingTravelDeltaGameLy(
-  heading: number,
-  travelLy: number
-): { dx: number; dy: number } {
-  const theta = (heading * Math.PI) / 180
-  return {
-    dx: travelLy * Math.sin(theta),
-    dy: travelLy * Math.cos(theta),
-  }
-}
+import { headingTravelDeltaGameLy } from './headingTravel'
 
 /** Per-turn ion storm travel delta in game light-years (heading 0 = north, clockwise). */
 export function ionStormStepDeltaGameLy(
